+++ include/ulsr/packet.h
@@ -40,17 +40,16 @@ struct ulsr_packet {
  * packets used for internal communication between devices in the simulation
  */
 enum ulsr_internal_packet_type {
-    PACKET_DATA, // data packets final destination will always be an external entity
+    PACKET_DATA = 0, // data packets final destination will always be an external entity
     PACKET_HELLO,
     PACKET_PURGE,
     PACKET_ROUTING,
     PACKET_ROUTING_DONE,
     PACKET_NONE,
+    PACKET_TYPE_COUNT,
 };
 
-static char *uslr_internal_type_to_str[] = {
-    "DATA", "HELLO", "PURGE", "ROUTING", "ROUTING_DONE", "NONE",
-};
+extern char *uslr_internal_type_to_str[PACKET_TYPE_COUNT];
 
 struct packet_route_t {
     u16 *path;
